AI Integrated Level Editor
Added ability to create design AI in the level editor, and levels can
now parse this information.

+++ DracosDescendentsLevelEditor/DracosDescendentsLevelEditor/EditDragonForm.cs
@@ -30,8 +30,8 @@ namespace DracosDescendentsLevelEditor
         {
             try
             {
-                dragon.x = Convert.ToSingle(xBox.Text);
-                dragon.y = Convert.ToSingle(yBox.Text);
+                dragon.x = Convert.ToInt32(xBox.Text);
+                dragon.y = Convert.ToInt32(yBox.Text);
             }
             catch
             {
